mesa: Put surfaceless last
I'm being a bit superstitious, as Mesa currently only cares what the
first one in the list is, treating the rest as an (unordered) set. But
that could change in the future, and I think it's just more intuitive to
list them in priority order. Mesa's own build system tends to put
surfaceless last fwiw too.