at-spi2-core: provide fallback dbus_daemon manually
This point directly to the NixOS current-system. This is necessary to
avoid creating a direct dependency on DBUS and increasing every app’s
runtime closure. Luckily, dbus_daemon is only used for the case when
dbus is not running and users can always run the dbus-daemon
themselves if it cannot be found.