xkbvalidate: Use $CC instead of hardcoded gcc
I initially didn't use $CC because I thought this would be GCC specific,
but it turns out that Clang actually accepts -std=gnu11.

So using $CC here might not work on compilers other than Clang or GCC,
but at the moment those are the compilers we typically use in nixpkgs,
so even if we'd use some other compiler it *might* even work there.

I've tested this by compiling against clangStdenv with both $CC and
clang hardcoded and it works.

This was reported by @dkudriavtsev on IRC.

Signed-off-by: aszlig <aszlig@nix.build>