iotop: use python3
The support of python2 will be dropped by the end of 2019, so python3
should be used by default here. In fact, this package supports both
python v2 and v3, however the installation with python3 broke because of a
bug in setup.py which is only fixed in the upstream's master[1].

Additionally, distros like debian already use iotop along with python3[2],
so this should be a sensible thing to do.

[1] https://repo.or.cz/iotop.git?a=commit;h=99c8d7cedce81f17b851954d94bfa73787300599
[2] https://packages.debian.org/de/buster/iotop