oracle-instantclient: 12.2.0.1.0 -> 19.3.0.0.0
It's now possible to download Oracle Instant Client without having to
click a checkbox to agree to the license, and then log in with an Oracle
account to start the download, at least on Linux (for now).

Darwin still seems to be at 18.1, and version numbers haven't been that
aligned recently, so allow different version numbers per platform.