netcdf4:init at 1.5.1.2
Generalized  the expression to be able to create a version with parallel
IO. If the function is called with netcdf-mpi, hdf5-mpi and mpi4py
it now produces a version that can access netcdf4 files in parallel.
I tested this with a small python script that uses parallel IO.
This should probably go to the original testsuite of the package which
is disabled in the nix expression I started with.
So I did not touch this part.
The patch is necessary since the original setup.py does not find
the parallel version of the netcdf although it is there since it
inspects the header file with a pattern that is propably only present in
the newest version. I made a pullrequest to the upstream maintainers to
generelize the pattern so that the patch will no longer be needed.