netcdffortran-mpi: new override
This is just a suggested override of netcdffortran wiht hdf5-mpi and netcdf-mpi as inputs.
The resulting fortran library can then be used for parallel IO.
Since I am new to nix it helped me a lot to find the netcdf-mpi package
directly before I was aware of the fact that it actually uses the same
function as netcdf just with the parallel version of hdf5. I therefore
suggest the same for the fortran binding.