lib: introduce lib.binPath
A lot of derivations have a *main* executable. For example pkgs.curl has
bin/curl, pkgs.bash has bin/bash. Writing those can be repetitive:

   "${lib.getBin pkgs.bash}/bin/bash"

Introduce a new `lib.binPath` function that can be used to infer a
default binary path for all packages.

   "${lib.binPath pkgs.bash}"

By default the binary is at $out/bin/$pname. For the package where that
doesn't match it's possible to provide a `meta.binPath` attribute that
contains the relative path to the binary.