{rustc,cargo}_1_36: add the older versions
I just copied the whole directory and reverted the bump of version+hash.
We don't seem likely to want to touch this again, so copying simplifies
future updates to rust.  I can't see why not keep the rest as in 1.37,
including the binary 1.36 for bootstrap.

Motivation: older firefox versions won't build with 1.37.