nixos-generate-config: Add system.copySystemConfiguration = true
This will be useful for the many people who accidentally delete their
configuration.nix file.

The alternative of changing the default of this option to true won't
work well because for some NixOS build, we can't assume for the
configuration file to be at the standard location. E.g. in nixops,
<nixos-config> will point to something completely different than what
the nixops evaluation thinks, therefore leaking the nixops deploy users
config into the deployment machines.