gnome2.libglade: make python2 build optional
Python2 is only needed for `libglade-convert`[1] which is a simple
script that converts old glade files into new glade files and for tests
and docs which we currently don't generate.

As Python2 is about to get EOLed and this is mostly a tool to migrate
old data, it shouldn't be built by default. With this change,
`xscreensaver` and `xsecurelock` don't depend on Python2 anymore.

[1] https://manpages.ubuntu.com/manpages/trusty/man1/libglade-convert.1.html