vault: 1.0.2 -> 1.2.2
This upgrades Vault to version 1.2.2. To accomplish this, we migrate to
using the `buildGoModule` helper, as since 1.0.2 the Vault build process
migrated to modules, and does not vendor its dependencies.

We also stop using the vault build script, and gox, as it only really
provides value for local development, where it configures GOOS/GOARCH
and installs into some convenient dev locations.