citrix_workspace: add 1908 (`citrix_workspace_19_8_0`)
New release:
https://www.citrix.de/downloads/workspace-app/linux/workspace-app-for-linux-latest.html
(unfortunately there's no version-specific link for the latest version).

Also added `preferLocalBuild = true;` to the derivation, due to
`requireFile` you have to build it yourself anyway, however I use
distributed builds by default and figured that this shouldn't be needed
since the longest part of the build would be the upload of the source
archive in that case.