Patch remind source to work on macOS
The default version of the source will show a banner on macOS and Windows which tells the user that the author would rather that they didn’t use remind on Apple or Microsoft products, and then sleep for 5 seconds.

There is no way for the user to remove this obstacle other than patch the source and recompile, which is also what other package managers do.

This commit also changes supported platform from linux to unix.