thinkpad/x1/6th-gen: fix trackpoint name
Newer ThinkPads have a new name for the Trackpoint - "TPPS/2 Elan TrackPoint".

Having the "[...] IBM Trackpoint" as default caused some weird
side-effects on my machine (e.g. stopped the trackpoint working after a
suspend) with the wrong name. Although this is mentioned in the option's
description[1], I think that this should be declared explicitly here.

As soon, as we supported newer models as well[2], we should probably
move this into its own common profile.

[1] https://nixos.org/nixos/options.html#hardware.trackpoint.device
[2] https://certification.ubuntu.com/catalog/component/input/5313/input%3ATPPS/2ElanTrackPoint/