nixos/tests/quake3: Fix evaluation error
In c814d72b517bb201c8bbbfc64e386c7023352886, a bunch of packages were
changed to use the pname attribute, among them were the quake3-demodata
and quake3-pointrelease which we use for the quake3 test.

Fortunately, having pname available means that we no longer need to
match using a prefix, so fixing this eval error also simplifies our
matching.

I directly pushed this to master because the change is non-controversial
and we can't break things that are already broken :-)

Signed-off-by: aszlig <aszlig@nix.build>