grub: enable copyKernels by default
When using filesystems such as ZFS, grub tries to directly access the
root drive to get the kernel. NixOS uses lots of hardlinks and grub
can have a problem reading that, resulting in messages such as:

  Error: external pointer tables not supported
  Error: you need to load the kernel first.

To avoid this issue, copy the kernel to /boot by default. This is how
it's done in gummiboot and probably other bootloaders.

Signed-off-by: William Casarin <jb55@jb55.com>