hadoop: fixing builds
Some annotation processing depends on tools.jar, which is usually
shipped with JDKs. Because we symlink JRE into the JDK, a traversal
from the JRE back to the JDK becomes impossible, so some assumptions
of relative paths break.

For ZHF: https://github.com/NixOS/nixpkgs/issues/68361

Tested build for 2.7, 3.0 and 3.1