qscintilla: Use both .so and _qt5.so suffix
Make sure that the both libqscintilla2_qt5.so and libqscintilla2.so are
present in `$out/lib` so all the dependant packages can find the most
appropriate library (some except the empty suffix, such as
sqlitebrowser, other expect the _qt5 suffix such as
python3Packages.qscintilla-qt5).

The mechanism would also work work for qt4 build, but it is broken at
the moment.

Also make sure python*Packages.qscintilla* are up to date