treewide: build some unfree packages locally
Unfree packages aren't distributed by our binary cache due to legal
reasons[1] and are usually a prebuilt binary that requires some patching.
When using distributed builds[2], those are uploaded to another build
machine as fixed-output derivations from `fetchurl` are built locally[3]
which takes a certain amount of time and resources with almost no gain
as the build process is trivial in contrast to the up/download to a
remote builder.

This is why I figured that at least some of the packages should be
explicitly built locally, I've done something simlar for
`citrix_workspace` already in the past[4].

The following packages are affected by this:

* `idea.*` (excluding free derivatives)
* `xmind`
* `teamviewer`

[1] https://nixos.wiki/wiki/FAQ/How_can_I_install_a_proprietary_or_unfree_package%3F#More_precision
[2] https://nixos.wiki/wiki/Distributed_build
[3] https://github.com/NixOS/nixpkgs/blob/267c8d6b2fea05bc811c8e2c2f4529b1436eeb9a/pkgs/build-support/fetchurl/default.nix#L95
[4] 87f818d9b21b16b067048a6ee0178b782886f0f0