nix-doc-tools: Produce an anchors index
What's the deal with the `cd`??

The filename used in the index are relative to cwd/the given path. This
means that to produce the expected filename associations, we need to be
in the right directory when producing the index.

Alternatively, the indexing tool could have been made to strip the
leading path from the filenames.