wlr-randr: init at unstable-2019-03-21
This is a simple clone of `xrandr` for wayland and is recommended by
sway's wiki[1] as replacement for `xrandr`.

Although the package is not stable yet, it's written by some sway
developers and appears to work pretty well.

When adding an additional monitor to your laptop (with the laptop being
"below" the monitor), this can be configured like this:

```
wlr-randr --output eDP-1 --on --pos 0,1080 --output HDMI-A-1 --on --pos 0,0
```

[1] https://github.com/swaywm/sway/wiki/i3-Migration-Guide#common-xorg-apps-used-on-i3-with-wayland-alternatives