ip2unix: 2.0.1 -> 2.0.2
This is just a small bugfix release (essentially adds two lines of code)
which fixes a segfault if using with a program that doesn't pass a
sockaddr buffer to accept() or accept4().

The reason I didn't cherry-pick this from master is because the version
there is version 2.1.x and I'd like to only backport fixes to NixOS
stable branches, so I released version 2.0.2 upstream specifically for
this.

Signed-off-by: aszlig <aszlig@nix.build>