ycmd: add gopls completer
ycmd gives 'no completer support' because of missing 'gopls'.
Add this as a conditional dependency.

Refactor the 'mkdir -p' step to be conditional per-dependency,
and placed just before the link step.
This is mostly for legibility but also pedantic correctness:
do not create a directory unless it will be used.

Signed-off-by: Sirio Balmelli <sirio@b-ad.ch>