systemd: don't use a custom repo anymore, but apply patches on top of systemd's sources
Keeping a custom branch and cherry-picking patches on top of vanilla
systemd has proven to be error-prone and hard to sync far too often.
We ship a chunk of patches in nixpkgs, and slowly move them to upstream
or fix inside NixOS where possible.

This patch simply inlines all the current commits on top of vanilla
systemd 243.