openscad: remove useless preBuild phase
The preBuild phase was introduced to follow a modification in qscintilla
build (76ad2796be28c5fbe8f8fe2b9bada12c3e1360e2) that forced the library
libqscintilla2_qt5.so to be called libqscintilla2.so. This has been
inproved in 4bf49d94179ad2378f50ba54138343fecb75e446 so the preBuild
trick is not necessary anymore.