build libinput to look for local quirks in /etc
If the user wants to configure libinput locally (e.g. to tune
touchpad sensitivity) the advertised[*] mechanism to do this
upstream is by adding a file /etc/libinput/local-overrides.quirks

Add --sysconfdir option to the libinput configurePhase so that we
build libinput to look for the file there instead of in libinput's
store path

[*] can't quite claim it as "supported", but it's certainly the
mechanism that's publically described.  See
https://wayland.freedesktop.org/libinput/doc/latest/device-quirks.html