mariadb: fix darwin build
We want to disable `PLUGIN_AUTH_PAM` when building:

1. `mariadb` on macOS.
2. `mariadb-client` on any platform

Unfortunately, the interaction of these two commits
  6c97b0486c97481dfdb82036de5382921269a771
  7e43b4d0aee408c3742b84e89749c2c494f16cfe
created a situation where we disable it *twice* when building on macOS.
Once in a darwin-specific `prePatch` script, and again in the `patches`
section for client builds.

This removes the redundant `prePatch` script and conditionally applies
the patch to `mariadb` server builds on darwin.

Fixes #70835