softether: Add vpnserver settings option
Allows to write vpnserver configuration in a nix map & translate it to
the native configuration file type.

Unfortunately there are some unmappable types like `byte` so the
solution to that is to interpret maps like:
```
{type: "byte", value: "whatever"}
```
as byte types.

Also add mutable option to specify either overwriting the config on
rebuild or keeping the original config around.